Updated Aoply Advanced ControlNet nodes to remove long unnecessaru model_optional fields, shifted some of the nodes.py code around to separate files to make deprecation cleaner
